feat(schedule): select all, Ctrl/Cmd+A, and hold Ctrl to show row checkboxes
- Select all is the first button in select mode. It selects every shift the
  grid draws, drafts and published, so a location filter narrows it. Row
  checkboxes then show checked, so one agent can be taken back out.
- Ctrl/Cmd+A does the same from anywhere on the schedule without entering
  select mode first. It stands down in inputs and dialogs, so text
  selection still works there.
- Holding Ctrl/Cmd for 200ms enters select mode. Releasing leaves it unless
  something was selected; then it stays, since a selection with no
  checkboxes on screen can't be seen or undone. Quick taps, Ctrl+shortcuts
  and Ctrl+wheel zoom don't trigger it, and blur or a hidden tab count as
  a release.

A publish or unpublish batch that got no answer is now reported as "not
confirmed" and followed by a refetch, rather than as failed: in the
incident this fixes, those shifts had in fact published.

+++ calendar-api-frontend/src/components/ScheduleCalendar/shift-dialogs/DuplicateDayDialog.tsx
@@ -152,7 +152,8 @@ const SourceDayPicker = ({
           <CalendarDays size={14} className="shrink-0 text-muted-foreground" />
           <span className="truncate text-[13px] font-semibold">
             {value.toLocaleDateString("en-US", {
-              weekday: "long",
+              // Short weekday: the long form truncated beside the chevron.
+              weekday: "short",
               month: "long",
               day: "numeric",
             })}
